pacer.js: isFilingAccount may not exist
If the `isFilingAccount` cookie does not exist, this raised
`TypeError: filingCookie is undefined`.

Test filingCookieVal before`.match()` is called on `undefined`.

Note that this cookie is our own invention, so if we're expecting it
to be set and it is not, it's because we failed to set it. Perhaps it
should be named differently to make provenance clear (also why is it a
cookie?).

xxx: This was broken by e2a07cf9 which refactored this code (why?).

Fixes this (new) test:

Chrome Headless 110.0.5478.0 (Mac OS 10.15.7) The PACER module hasFilingCookie returns false for a valid logged in cookie without isFilingAccount FAILED
        TypeError: Cannot read properties of undefined (reading 'match')
            at Object.hasFilingCookie (src/pacer.js:605:27 <- src/pacer.js:265:303)
            at UserContext.<anonymous> (spec/PacerSpec.js:587:20)
            at <Jasmine>