recap.js: Call callbacks on failure. XXX
Where `callback()` functions are usually `cb()`,
generally invoke `callback(false)` upon failure, rather than doing nothing.
Callbacks are consistently invoked as `cb(data || null)` upon success.

XXX: IS IT NOT CLEAR THIS IS A GOOD IDEA.
It may help with debugging.

But callbacks are kind of overloaded, and a successful callback that returns a falsy value (like null,
for emptiness), so this is kind of confusing and it may be hard to understand the implications.