feat(appellate): Tighten the criteria in the wrapperMutationObserver
Tightens the criteria in wrapperMutationObserver to prevent unnecessary re-renders and improve performance. Previously, the callback would run the logic to remove the event listener from the button and gather the data to upload the PDF whenever any div element was added to the page, even if irrelevant to the extension. This could lead to performance problems, especially when adding banners.

The new criteria ensures that the callback function will run the logic only when the mutation that adds content to the Download Confirmation page is included in the array of MutationRecord objects.