feat(appellate): Tweaks the handleAcmsDocket method to check the footer
Introduces a mechanism to verify that the page has fully loaded when updating RECAP icons. This effectively addresses the issue with RECAP links that occurred when users clicked the 'Refresh RECAP links' option in the RECAP button. By ensuring that the icons are refreshed even when the page is fully rendered, this improvement eliminates the need for a complete page reload.